Added barrelling for progress bar
